% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllSigTest.R
\name{AllSigTest}
\alias{AllSigTest}
\title{Univariate & Bivariate analysis of dataset}
\usage{
AllSigTest(Data, Targetx)
}
\arguments{
\item{Input}{data frame}
}
\value{
charts
}
\description{
This function loads a data frame, provides univariate distribution of each variable
and performs significance tests on a target variable with all other variables. 
The tests are correlation, chi-square and ANOVA depending no the type of varible.
}

